Flytt funksjoner inn i testmappen, de skal slettes
