+++ src/komponenter/kalkulator/utils/beregningerForAGP2.ts
@@ -21,91 +21,6 @@ export enum Permitteringssituasjon {
     AGP2_IKKE_NÅDD_PGA_IKKE_PERMITTERT_VED_INNFØRINGSDATO = 'AGP2_IKKE_NÅDD_PGA_IKKE_PERMITTERT_VED_INNFØRINGSDATO',
 }
 
-export interface InformasjonOmAGP2Status {
-    type: Permitteringssituasjon;
-
-    sluttDato: Dayjs | undefined;
-
-    gjenståendePermitteringsdager: number;
-    bruktePermitteringsdager?: number;
-
-    brukteDagerVedInnføringsdato: number;
-    fraværsdagerVedInnføringsdato: number;
-    permitteringsdagerVedInnføringsdato: number;
-    permittertVedInnføringsdato?: boolean;
-
-    finnesLøpendePermittering?: boolean;
-}
-
-export const finnInformasjonAGP2 = (
-    tidslinje: DatoMedKategori[],
-    innføringsdatoAGP2: Dayjs,
-    erLøpende: boolean,
-    dagensDato: Dayjs,
-    antallDagerFørAGP2Inntreffer: number,
-    finnesLøpende: boolean
-): InformasjonOmAGP2Status => {
-    const situasjon = finnPermitteringssituasjon(
-        tidslinje,
-        innføringsdatoAGP2,
-        antallDagerFørAGP2Inntreffer
-    );
-    const oversiktOverPermitteringVedInnføringsdato = finnOversiktOverPermitteringOgFraværGitt18mnd(
-        innføringsdatoAGP2,
-        tidslinje
-    );
-
-    const dataVedInnføringsdato = {
-        fraværsdagerVedInnføringsdato:
-        oversiktOverPermitteringVedInnføringsdato.dagerAnnetFravær,
-        permitteringsdagerVedInnføringsdato:
-        oversiktOverPermitteringVedInnføringsdato.dagerPermittert,
-        brukteDagerVedInnføringsdato:
-        oversiktOverPermitteringVedInnføringsdato.dagerBrukt,
-    };
-
-    let dataSpesifikkForSituasjon;
-    switch (situasjon) {
-        case Permitteringssituasjon.AGP2_NÅDD_VED_INNFØRINGSDATO:
-            dataSpesifikkForSituasjon = {
-                sluttDato: innføringsdatoAGP2,
-                gjenståendePermitteringsdager: 0,
-                bruktePermitteringsdager: antallDagerFørAGP2Inntreffer,
-                permittertVedInnføringsdato: true,
-            };
-            break;
-        case Permitteringssituasjon.AGP2_NÅDD_ETTER_INNFØRINGSDATO:
-            dataSpesifikkForSituasjon = getInformasjonOmAGP2HvisDenNåsEtterInnføringsdato(
-                tidslinje,
-                innføringsdatoAGP2,
-                antallDagerFørAGP2Inntreffer
-            );
-            break;
-        case Permitteringssituasjon.AGP2_IKKE_NÅDD_PGA_IKKE_PERMITTERT_VED_INNFØRINGSDATO:
-            dataSpesifikkForSituasjon = {
-                sluttDato: innføringsdatoAGP2,
-                gjenståendePermitteringsdager: 0,
-                permittertVedInnføringsdato: false,
-            };
-            break;
-        case Permitteringssituasjon.AGP2_IKKE_NÅDD_PGA_FOR_LITE_PERMITTERT:
-            dataSpesifikkForSituasjon = getInformasjonOmAGP2HvisAGP2IkkeNås(
-                tidslinje,
-                innføringsdatoAGP2,
-                antallDagerFørAGP2Inntreffer,
-                dagensDato
-            );
-            break;
-    }
-
-    return {
-        type: situasjon,
-        finnesLøpendePermittering: finnesLøpende,
-        ...dataVedInnføringsdato,
-        ...dataSpesifikkForSituasjon,
-    };
-};
-
 export const finnPermitteringssituasjon = (
     tidslinje: DatoMedKategori[],
     innføringsdatoAGP2: Dayjs,
